Move to using name as the primary key used to construct @ids
name should be a short url friendly string, display_name is used
as the longer human friendly name.

This also finishes the move from url to @id, and to get both
accessible I need to have both @id and name in the serializers.

For the mouse strains I decided to use what was previous the CODE as
the mouse strain name and add a new field for the id that the DACC
wants.