so many updates to importer notebook
Tried reading from ods files, but discovered that comments end up in
the cell values. There's some code trying to suppress that, but it
wasn't worth it.

Some of the pituitary samples got reclassified as diencephalon, as the
dissection might not be perfect.

The F1 submission spreadsheets were added to the notebook

A bit of fastq parsing got moved out of the notebook and into
igvf_mice.io.read_fastq_metadata

And there's some of my efforts to deal with the IGVF_EX2 plate where
sublibraries link to different plates, but the changes didn't work,
and now I'm committing the current progress so I can try a more
intrusive change to moving plate into libraryinrun